feat(core): Rework User/auth implementation to enable 3rd party auth
Relates to #215. In this commit we are adding a new AuthenticationMethod entity, which stores authentication information for a given auth provider. The default username/password strategy is known as the "Native" strategy.

This refactor does not add any new functionality, but sets up support for external auth providers whilst making sure all existing e2e tests pass.

BREAKING CHANGE: A new `AuthenticationMethod` entity has been added, with a one-to-many relation to the existing User entities. Several properties that were formerly part of the User entity have now moved to the `AuthenticationMethod` entity. Upgrading with therefore require a careful database migration to ensure that no data is lost. On release, a migration script will be provided for this.